[Bronze V] Title: Adding Trouble, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/31654. Adding Trouble/Adding Trouble.cc
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    long long a, b, c;
+    int a, b, c;
     cin >> a >> b >> c;
     
     if (a + b == c) cout << "correct!";
